+++ src/main.py
@@ -39,7 +39,7 @@ import queue
 from tkinter import filedialog
 from drivers.factory import AutoDetectDriver
 
-APP_VERSION = "2.6.1"
+APP_VERSION = "3.0.0"
 
 
 class ToolTip: